[svn r560] Convert derby test benchmark over to using the TPC-C like code contained in derby.
This benchmark wraps the operations of the TPC-C like tests specified in
  org.apache.derbyTesting.system.oe.direct.Standard
  to allow transactions that fail (normally due to deadlock or timeouts)
  to be retried.

At the end of an iteration the transaction mix is specified as a sanity check
which gets digested.

Configuration allows the setting of the key TPC-C controls:
    scale, number of ware houses, number of terminals and the number of
    transactions
It also allows the setting of the timeouts for deadlocks and transactions
as well as the retry limit.  Note that deadlock timeout setting has a
significant impact on the performance as there seems to be a lot of
contention.

NOTE: The benchmark uses derby-10.5.3.0, however, it has also been
successfully tried with derby code trunk with revision 824955 so it
should be easily adaptible to 10.6.1.0 when it is released.