fix some minor errors.
There is a problem with classpath such that javadoc wasn't able to find
`commons.cli` package;
There are also small problems with using angle brackets `<` & `>` that
caused javadoc to complain 'malformed HTML tag'.

After this fix, the remaining errors are all about `@date` and `@id`
tags. These are not supported by javadoc. However they don't pose a
problem to the generated HTML document. So these problems are ignored
for now.