WIP: PMD 5.7.0 now compiles
PMD 5.7.0 use Maven as its build system. The `build.xml` for the
benchmark has been changed to call Maven.
Also, it requires Java JDK toolchain to be configured for Maven. A
shell script has been added to generate the toolchain.xml file.

Note that during the building process, PMD requires both 1.7 and 1.8
toolchain to be configured. A clean solution would require the JDK7 to
be installed on the platfor along with JDK8/9. This requirement can be
annoying. Therefore the generated toolchain.xml file
has a hack such that if the platform java version is >= 1.8, it will
generate two entries (1.7 and the actual version) that point to the same
JDK directory.

Also the download link for the release has been changed to GitHub.