Rerote PMD harness
PMD 5.7.0 supports passing in file lists; thus `prepended_filelist()` is
added to generate a new file list with scratch directory prepended to
each entry in the list. The original file list is renamed with an
`.orig` suffix.

`prepare()` is modified to conform to the new commandline interface. In
particular, flags are added and language and version are specified.

Also sets a system property so that PMD doesn't just exit after run.