[batik, fop] remove fop's interference with batik causing it to throw exception
Fop carries a batik-all-1.9.jar with it, which will overwrite the batik benchmark jar
when unzipped into scratch directory, rendering the patch to batik 1.9 ineffective.
Fop benchmark doesn't seem to actually use the jar. So removing it from the list of
jars to include in fop doesn't affect running fop benchmark, and batik is now fine.