fix: fop's dependency on batik
If batik is built before fop, fop's lib dependency on batik
will overwrite the patched batik jar.
There was a previous commit that attempts to deal with the issue,
but it created dependency not found problem when
fop is the only one built.
This commit properly deals with the issue by renaming the
batik jar that fop depends on, thus not overwriting the patched
batik benchmark.