fix: use script command to enable tty to run gitlab-runner locally
When trying to locally test the CI problem, there was another
issue with running gitlab-runner locally.
It seems that during the deploy stage of building DayTrader,
Geronimo tries to execute `stty -icanon min 1 < /dev/tty` command.
It causes the whole thing to hang.
The reason might be that tty is not enabled in gitlab-runner.
I found a hacky
[solution ](https://gitlab.com/gitlab-org/gitlab-runner/issues/1021)
that uses the script command to emulate tty. And it seems to work.