fix: DayTrader GitLab CI problem (Issue #16)
The problem was because Geronimo server wants to read from stdin.
GitLab runner doesn't provide stdin, so it just reads an EOF and quits.
This is why GBeans is not running.
Thus the solution is to keep stdin open for the duration of the benchmark.
Setting a time out (120 seconds) with `sleep` command seems to work.
NOTE: this value may need to be modified on slower machines.